abandoned splitting the behavior layer into sub-modules
